Refactor all 3 to 1 generic method to find maximum

+++ Program.cs
@@ -14,7 +14,6 @@ namespace TestMaximum
             TestMax maximum = new TestMax();
             Console.WriteLine("Select from given  below.....\n1. Max in Three Integers \n2 Max in Three Floats \n3. Max in string");
             int option = Convert.ToInt32(Console.ReadLine());
-           
             switch (option)
             {
                 case 1:
@@ -39,7 +38,7 @@ namespace TestMaximum
                     float floatNum2 = float.Parse(Console.ReadLine());
                     Console.WriteLine("Enter third Float value");
                     float floatNum3 = float.Parse(Console.ReadLine());
-                    float floatMax = maximum.MaxInThree((int)floatNum1, (int)floatNum2, (int)floatNum3);
+                    float floatMax = maximum.MaxInThree(floatNum1, floatNum2, floatNum3);
                     if (floatMax == 0)
                     {
                         Console.WriteLine("Two or Three numbers have equal value");
@@ -56,12 +55,13 @@ namespace TestMaximum
                     string str3 = Console.ReadLine();
                     string strMax = maximum.MaxInThree(str1, str2, str3);
                     if (strMax == null)
-                    { 
+                    {
                         Console.WriteLine("Two or Three strings are same");
                     }
                     else
                         Console.WriteLine("Maximum out of given strings {0} {1} & {2} is {3}", str1, str2, str3, strMax);
                     break;
+
                 default:
                     Console.WriteLine("Invalid Input");
                     break;
